feature: moving to right path
