+++ example/src/main/java/com/mercadopago/sdk/android/example/di/PresentationModule.kt
@@ -3,6 +3,7 @@ package com.mercadopago.sdk.android.example.di
 import com.mercadopago.sdk.android.example.MainApplication
 import com.mercadopago.sdk.android.example.presentation.coremethods.PaymentScreenViewModel
 import com.mercadopago.sdk.android.example.presentation.logs.LogsViewModel
+import com.mercadopago.sdk.android.example.presentation.mpextended.MPExtendedViewModel
 import com.mercadopago.sdk.android.example.presentation.sdkinitializer.SdkInitializerViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
@@ -15,4 +16,5 @@ internal fun providePresentationModule(application: MainApplication) = module {
         )
     }
     viewModel { LogsViewModel() }
+    viewModel { MPExtendedViewModel() }
 }
