+++ mp-extended/src/main/java/com/mercadopago/sdk/android/mpextended/data/datasource/remote/MPExtendedRemoteDataSource.kt
@@ -1,9 +1,9 @@
-package com.mercadopago.sdk.android.coremethods.builder.data.datasource.remote
+package com.mercadopago.sdk.android.mpextended.data.datasource.remote
 
-import com.mercadopago.sdk.android.coremethods.builder.data.remote.request.MPDeviceSessionIdRequest
-import com.mercadopago.sdk.android.coremethods.builder.domain.model.MPDeviceSession
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
+import com.mercadopago.sdk.android.mpextended.data.remote.request.MPDeviceSessionIdRequest
+import com.mercadopago.sdk.android.mpextended.domain.model.MPDeviceSession
 
 internal interface MPExtendedRemoteDataSource {
     suspend fun getDeviceSessionId(
