+++ mp-extended/src/main/java/com/mercadopago/sdk/android/mpextended/di/MPExtendedModulesProvider.kt
@@ -1,12 +1,12 @@
-package com.mercadopago.sdk.android.coremethods.builder.di
+package com.mercadopago.sdk.android.mpextended.di
 
 import com.mercadopago.sdk.android.core.di.CoreKoinFactory
 import com.mercadopago.sdk.android.core.di.CoreKoinModuleProvider
-import com.mercadopago.sdk.android.coremethods.builder.di.datasource.provideDataSourceModule
-import com.mercadopago.sdk.android.coremethods.builder.di.repository.provideRepositoryModule
-import com.mercadopago.sdk.android.coremethods.builder.di.services.provideNetworkModule
-import com.mercadopago.sdk.android.coremethods.builder.di.usecases.provideUseCaseModule
 import com.mercadopago.sdk.android.di.MercadoPagoKoinComponent
+import com.mercadopago.sdk.android.mpextended.di.datasource.provideDataSourceModule
+import com.mercadopago.sdk.android.mpextended.di.repository.provideRepositoryModule
+import com.mercadopago.sdk.android.mpextended.di.services.provideNetworkModule
+import com.mercadopago.sdk.android.mpextended.di.usecases.provideUseCaseModule
 import org.koin.core.Koin
 import org.koin.core.module.Module
