+++ mp-extended/src/main/java/com/mercadopago/sdk/android/mpextended/di/services/NetworkModule.kt
@@ -1,7 +1,7 @@
-package com.mercadopago.sdk.android.coremethods.builder.di.services
+package com.mercadopago.sdk.android.mpextended.di.services
 
 import com.mercadopago.sdk.android.core.di.RetrofitServiceFactory
-import com.mercadopago.sdk.android.coremethods.builder.data.remote.service.MPExtendedService
+import com.mercadopago.sdk.android.mpextended.data.remote.service.MPExtendedService
 import org.koin.core.module.Module
 import org.koin.dsl.module
