+++ mp-extended/src/main/java/com/mercadopago/sdk/android/mpextended/domain/repository/MPExtendedRepository.kt
@@ -1,9 +1,9 @@
-package com.mercadopago.sdk.android.coremethods.builder.domain.repository
+package com.mercadopago.sdk.android.mpextended.domain.repository
 
-import com.mercadopago.sdk.android.coremethods.builder.domain.model.MPDeviceSession
-import com.mercadopago.sdk.android.coremethods.builder.domain.model.params.GetDeviceSessionParams
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
+import com.mercadopago.sdk.android.mpextended.domain.model.MPDeviceSession
+import com.mercadopago.sdk.android.mpextended.domain.model.params.GetDeviceSessionParams
 
 internal interface MPExtendedRepository {
     suspend fun getDeviceSession(
